Scoped policies for server groups
Server groups are currently used to ensure that instances in the
groups are either grouped together on the same host (affinity) or
spread across different hosts (anti-affinity).  This
affinity/anti-affinity concept could be broadly useful for a number of
use cases, but many of these use cases would require (anti-)affinity
with a scope broader than a single host.  This spec adds a general and
flexible scoping mechanism to server group policies which enables
these use cases.  Under this mechanism, "host" is just one of
(potentially) many available scopes; by making host the default scope
if none is specified, this mechanism is fully backward-compatible with
the existing server group policies.  Arbitrary additional scopes can
be defined by the deployer.

Change-Id: I6d121bf71c949d2349671b94464f072099cfb795