Add update operations to server groups
Right now it is not possible to change a server group's policies after
the group is created, nor is it possible to remove instances from a
group, nor to add an instance to a group after boot time.  This spec
adds these operations and clarifies the semantics of existing
operations (like migrations) with respect to server group policies.

blueprint: server-groups-support-update
ApiImpact

Change-Id: Ic0ad8270c0b983b75e3b2511f47863451e658223